Migrated to SDK 15.3.0

+++ ble_cus.h
@@ -26,7 +26,7 @@ NRF_SDH_BLE_OBSERVER(_name ## _obs,                                             
 
 #define CUSTOM_SERVICE_UUID               0x1400
 #define CUSTOM_VALUE_CHAR_UUID            0x1401
-																					
+
 /**@brief Custom Service event type. */
 typedef enum
 {
@@ -54,7 +54,7 @@ typedef void (*ble_cus_evt_handler_t) (ble_cus_t * p_bas, ble_cus_evt_t * p_evt)
 typedef struct
 {
     ble_cus_evt_handler_t         evt_handler;                    /**< Event handler to be called for handling events in the Custom Service. */
-    uint8_t                       initial_custom_value;           /**< Initial custom value */
+    uint16_t                       initial_custom_value;           /**< Initial custom value */
     ble_srv_cccd_security_mode_t  custom_value_char_attr_md;     /**< Initial security level for Custom characteristics attribute */
 } ble_cus_init_t;
 
@@ -65,7 +65,7 @@ struct ble_cus_s
     uint16_t                      service_handle;                 /**< Handle of Custom Service (as provided by the BLE stack). */
     ble_gatts_char_handles_t      custom_value_handles;           /**< Handles related to the Custom Value characteristic. */
     uint16_t                      conn_handle;                    /**< Handle of the current connection (as provided by the BLE stack, is BLE_CONN_HANDLE_INVALID if not in a connection). */
-    uint8_t                       uuid_type; 
+    uint8_t                       uuid_type;
 };
 
 /**@brief Function for initializing the Custom Service.
@@ -83,7 +83,7 @@ uint32_t ble_cus_init(ble_cus_t * p_cus, const ble_cus_init_t * p_cus_init);
  *
  * @details Handles all events from the BLE stack of interest to the Battery Service.
  *
- * @note 
+ * @note
  *
  * @param[in]   p_cus      Custom Service structure.
  * @param[in]   p_ble_evt  Event received from the BLE stack.
@@ -95,10 +95,10 @@ void ble_cus_on_ble_evt( ble_evt_t const * p_ble_evt, void * p_context);
  * @details The application calls this function when the cutom value should be updated. If
  *          notification has been enabled, the custom value characteristic is sent to the client.
  *
- * @note 
- *       
+ * @note
+ *
  * @param[in]   p_bas          Custom Service structure.
- * @param[in]   Custom value 
+ * @param[in]   Custom value
  *
  * @return      NRF_SUCCESS on success, otherwise an error code.
  */
